redacting functions

+++ board.c
@@ -26,7 +26,7 @@ typedef struct{
 typedef struct{
     small_board boards[3][3];
     players_turn current_player;
-    size_t next_row, next_col;
+    int next_row, next_col;
 }big_board;
 
 void init(big_board * board) { //initting board with EMPTY type
